Have tests return int instead of size_t, for Tinker update.

+++ src/str/charsplit_test.c
@@ -6,7 +6,7 @@
 
 #include "charsplit_test.h"
 
-size_t test_ali_str_charsplit()
+int test_ali_str_charsplit()
 {
     TEST_HAS_ASSERTIONS();
